+++ tasks/repo/mysql/structs.go
@@ -24,7 +24,7 @@ type dbTask struct {
 	CategoryID   uint64        `db:"category_id"`
 	ProjectID    sql.NullInt64 `db:"project_id"`
 	Priority     bool          `db:"priority"`
-	ExecOrder    uint64        `db:"exec_order"`
+	ExecOrder    int           `db:"exec_order"`
 }
 
 type dbTaskForList struct {
@@ -35,7 +35,7 @@ type dbTaskForList struct {
 	InWorkTime   time.Duration `db:"in_work_time"`
 	StatusID     uint64        `db:"status_id"`
 	Priority     bool          `db:"priority"`
-	ExecOrder    uint64        `db:"exec_order"`
+	ExecOrder    int           `db:"exec_order"`
 }
 
 type dbCategory struct {
